Refactor transaction handling and sequence generation
Removed the rwlock package and refactored the transaction and sequence handling by introducing a sequence number generator and submit queue. Transactions are now submitted but not monitored for inclusion in a block on chain (for speed). Confirming transaction success is left up to the caller.

This is done to improve throughput and avoid sequence number collisions on chain when submitting multiple transactions for inclusion for a given block. It assumes this thread is the only entity submitting transactions using the passed private key.

closes #5