chore: manually set account sequence number on creating AA (#9)
* Update dev script in package.json

The development script in package.json has been updated. Previously, it was simply running nodemon on the src/app.ts file. Now, it has been modified to run nodemon with dotenv, which allows environment variables to be set before executing the script.

* Refactor transaction handling and sequence generation

Refactored the transaction and sequence handling by introducing a sequence number generator and submit queue. Transactions are now submitted but not monitored for inclusion in a block on chain (for speed). Confirming transaction success is left up to the caller.

This is done to improve throughput and avoid sequence number collisions on chain when submitting multiple transactions for inclusion for a given block. It assumes this thread is the only entity submitting transactions using the passed private key.

closes #5

---------

Co-authored-by: justin <328965+justinbarry@users.noreply.github.com>